requirements.txt
# Core (existing project)
opencv-python
pytesseract
google-generativeai
python-dotenv
pydantic>=2.0
requests>=2.28.0

# Pipeline architecture
redis>=5.0
SQLAlchemy>=2.0
fastapi>=0.110
uvicorn[standard]>=0.27
